[Decomp][Spell] Cohesion split: extract skill/enchant/pet handlers to SpellEffectSkillEnchantPet.cpp
Move the contiguous learn/skill/enchant/pet/weapon block (EffectLearnSpell
through EffectWeaponDmg) out of SpellEffects.cpp into a new
SpellEffectSkillEnchantPet.cpp translation unit.

Pure file-level cohesion split: declarations stay in Spell.h, the
SpellEffects[] dispatch table stays in SpellEffects.cpp and resolves the
moved member methods by address at link time. Byte-identical move, no
behaviour change. SpellEffects.cpp 6027 -> 4706.
